backend/requirements.txt
fastapi
uvicorn[standard]
pyodbc
pydantic 
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4